making the site safe

+++ blog/views.py
@@ -27,28 +27,42 @@ def detail(request, blog_id):
 
 @login_required
 def delete(request, blog_id):
-    post = get_object_or_404(Blog, pk=blog_id)
-    post.delete()
-    messages.success(request, 'Post has been deleted')
-    return redirect('index')
+    if request.user.is_superuser:
+        post = get_object_or_404(Blog, pk=blog_id)
+        post.delete()
+        messages.success(request, 'Post has been deleted')
+        return redirect('index')
+    else:
+        messages.success(request, 'You are not authorized')
+        return redirect('index')
 
 
 @login_required
 def update(request, blog_id, template_name='main/form.html'):
-    post = get_object_or_404(Blog, pk=blog_id)
-    form = BlogForm(request.POST or None, instance=post)
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Post has been updated')
+    if request.user.is_superuser:
+        post = get_object_or_404(Blog, pk=blog_id)
+        form = BlogForm(request.POST or None, instance=post)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Post has been updated')
+            return redirect('index')
+        return render(request, template_name, {'form': form})
+    else:
+        messages.success(request, 'You are not authorized')
         return redirect('index')
-    return render(request, template_name, {'form': form})
 
 
 @login_required
 def create(request, template_name='main/form.html'):
-    form = BlogForm(request.POST or None)
-    if form.is_valid():
-        form.save()
-        messages.success(request, 'Post has been created')
+    if request.user.is_superuser:
+        form = BlogForm(request.POST or None)
+        if form.is_valid():
+            form.save()
+            messages.success(request, 'Post has been created')
+            return redirect('index')
+        return render(request, template_name, {'form': form})
+    else:
+        messages.success(request, 'You are not authorized')
         return redirect('index')
-    return render(request, template_name, {'form': form})
+
+
